Work around exceptions being silently omitted in Application Insights due to long messages
It's this issue I believe: https://github.com/microsoft/ApplicationInsights-Java/issues/482
Was logging the page source as a way to see if what we're scraping lines up with what a normal user gets when browsing to the same page. Would still be useful but probably not crucial, so just remove for now.